split ias15.py into a subpackage structure
