+++ src/jorbit/mpchecker/ephem_generation/README.md
@@ -1,1 +1,4 @@
+**NOTE:**
+Do not trust/use scripts in this folder. We will not be generating another cached ephemeris until v2.0, and when we do it will be with an entirely new pipeline. The scripts here are no longer maintained and may not even run. They are only meant to provide a rough sketch of how the ephemeris was generated in the first place, and to guide setup when we do it again in the future.
+
 Files in this folder contain snippets of code ran to generate the Chebyshev coefficients describing the motion of objects in the MPC's MPCORB.DAT file. They are not intended to be run as standalone scripts and are only meant to guide setup when updating the ephemeris sometime in the future. They are also not stiched into a coherenet pipeline, charitably because I assume different clusters will require different setups and uncharitably because I was sloppy when creating the first ephemeris. This process requires a computing cluster, and the last step (as written) requires holding the entire ephemeris in memory.
